+++ src-docs/docs/changelog.md
@@ -2,6 +2,13 @@
 
 This is the list of what has changed since the first public version of *DuRPresets*.
 
+## 0.0.3
+
+### Improvements
+
+- Added an option to ignore output filename when importing. This is now the default.
+- Added support for playblast options from *DuBlast*.
+
 ## 0.0.2
 
 ### Fixes